Passing 1202 failing 44/1948 MP RS operators tck tests

+++ microprofile/reactive-streams/src/test/java/io/helidon/microrofile/reactive/CountingSubscriber.java
@@ -83,7 +83,7 @@ public class CountingSubscriber implements Subscriber<Long> {
     }
 
     public void expectRequestCount(int n) {
-        assertEquals(n, requestCount.get(), String.format("Expected %d requests but only %d received.", n, requestCount.get()));
+        assertEquals(n, requestCount.get(), String.format("Expected %d requests but only %d received.", n, (long)requestCount.get()));
     }
 
     public void expectSum(long n) {
